+++ Modules/Shop/Resources/assets/js/components/ErrorAlert.vue
@@ -1,9 +1,6 @@
 <template>
-    <div class="alert alert-danger alert-dismissible fade show" role="alert">
+    <div class="alert alert-danger" role="alert">
         <icon name="exclamation-triangle"></icon> {{ message }}
-        <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-            <span aria-hidden="true">&times;</span>
-        </button>
     </div>
 </template>