+++ Modules/Shop/Routes/api.php
@@ -23,5 +23,7 @@ Route::middleware(['auth', 'language'])
             Route::get('card', 'ShopController@getCard')->name('getCard');
             Route::patch('card/{handout}/redeem', 'ShopController@redeemCard')->name('redeemCard');
             Route::delete('card/{handout}', 'ShopController@cancelCard')->name('cancelCard');
+            Route::get('/summary', 'ShopController@summary')->name('summary');
+            Route::post('/deleteNonRedeemed', 'ShopController@deleteNonRedeemed')->name('deleteNonRedeemed');
         });
     });