+++ Changelog.md
@@ -3,6 +3,7 @@
 ## 2.10.1
 
 * Added support for QR code scanner functioning when camera is not available.
+* Added shop card manager to remove non-redeemed cards
 
 ## 2.10.0